refactor: GameController 생성하고 이 곳에서 InputView 사용하도록 함
